Fix cyclic dependencies
The LispMNHelper depends on the internet-apps module, but this
cannot be declared as waf refuse to compile, if done.
This creates linking errors if compiling scripts that are including
the internet module but not the internet-apps module.
We commented the LispMNHelper as a quick fix.
It should be moved to an other module for a long term fix.